update reuse input

+++ src/app/auth/login/login.component.ts
@@ -1,9 +1,9 @@
-import { User } from 'src/app/models/user';
 import { UserService } from './../../Services/user.service';
-import { FormBuilder, Validators } from '@angular/forms';
+import { FormBuilder } from '@angular/forms';
 import { Component } from '@angular/core';
 import { MessageService } from 'primeng/api';
 import { Router } from '@angular/router';
+import { validateEmail, validateRequired } from 'src/app/validators';
 
 @Component({
   selector: 'app-login',
@@ -22,31 +22,41 @@ export class LoginComponent {
 
   error!: any;
 
-  loginForm = this.fb.group({
+  required = validateRequired;
+  email = validateEmail;
+
+  loginForm: any = this.fb.group({
     email: [''],
     password: [''],
   });
 
   click() {
-    this.userService.loginUser(this.loginForm.value).subscribe(
-      (data) => {
-        localStorage.clear();
-        this.token = data;
-        localStorage.setItem('token', this.token.token);
-        localStorage.setItem('id', this.token.id);
-        localStorage.setItem('name', this.token.name);
-        this.name.next(this.token.name);
-        this.router.navigate(['/']);
-      },
-      (error) => {
-        this.messageService.add({
-          severity: 'error',
-          summary: 'Error',
-          detail: error,
-        });
-        this.error = error.error.errors;
-        console.log(this.error);
-      }
-    );
+    if (!this.loginForm.invalid) {
+      this.userService.loginUser(this.loginForm.value).subscribe(
+        (data) => {
+          localStorage.clear();
+          this.token = data;
+          localStorage.setItem('token', this.token.token);
+          localStorage.setItem('id', this.token.id);
+          localStorage.setItem('name', this.token.name);
+          this.name.next(this.token.name);
+          this.router.navigate(['/']);
+        },
+        (error) => {
+          this.messageService.add({
+            severity: 'error',
+            summary: 'Error',
+            detail: error,
+          });
+          this.error = error.error.errors;
+        }
+      );
+    } else {
+      this.messageService.add({
+        severity: 'error',
+        summary: 'Error',
+        detail: 'Please fill all the fields',
+      });
+    }
   }
 }
